+++ internal/middleware/logger.go
@@ -1,0 +1,31 @@
+package middleware
+
+import (
+	"net/http"
+	"time"
+
+	"go.uber.org/zap"
+)
+
+var sugar zap.SugaredLogger
+
+func WithLogging(h http.Handler) http.Handler {
+	logFn := func(w http.ResponseWriter, r *http.Request) {
+		start := time.Now()
+
+		uri := r.RequestURI
+		method := r.Method
+
+		h.ServeHTTP(w, r) // обслуживание оригинального запроса
+
+		duration := time.Since(start)
+
+		sugar.Infoln(
+			"uri", uri,
+			"method", method,
+			"duration", duration,
+		)
+
+	}
+	return http.HandlerFunc(logFn)
+}
